moved large_example to examples/ and updated BUILD files
